first modifications
git-svn-id: https://support.ginsys.be/svn/lt/trunk@42 ff10d734-6443-0410-b6ef-c23fcb60535c

+++ make.cmd
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+if [ ! -e ../static ]
+        then mkdir ../static && bin/setup.cmd ; exit
+fi
+
 # cleanup
 rm -rf book*
